Implemented the Task Object
Created the Task Model, Service, Controller, and Dto. I created the TaskAllowedDto in case we wanted to make sure that the member could only access the tasks that they were assigned to them by the admin. Similar to role permissions, it would allow admin to see who was assigned to which task.