updated subsystems + new buttons for user interaction
I updated the info in the dummy data to include the new software system. I also created the Delete button that (for now) temporarily deletes a member and their data until the user refreshes the page. I also created an option for the user to navigate to the individual gallery card view when they click on a row in the MembersTable. I also fixed the filtering logic so that the filters persist without the filter dropdown needing to be called first by moving the consts to the parent function instead of them living in the child function.